Consolidate image and cube I/O in custom_io.py
